fix: 동적 라우팅, 옵셔널 세그먼트로 변경
