+++ pages/folder/[[...folderId]].tsx
@@ -12,6 +12,7 @@ import AddModal from '../../components/Modal/AddModal/AddModal';
 import ModalPortal from '../../Portal';
 import { UserContext } from '@/contexts/UserContext';
 import FolderModals from '@/components/FolderModalContainer/FolderModals';
+import { useRouter } from 'next/router';
 
 function Folder() {
   const id = useContext(UserContext);
@@ -22,8 +23,11 @@ function Folder() {
   const [searchKeyword, setSearchKeyWord] = useState('');
   const [url, setUrl] = useState('');
   const [toggleInput, setToggleInput] = useState(true);
-  const { linkList, loading } = useGetFolder(id, searchKeyword, onSelect.id);
-  const { link } = useGetFolderList(id);
+  const [wrongFolder, setWrongFolder] = useState(false);
+  const router = useRouter();
+  const folderId = router.query.folderId as string;
+  const { linkList, loading } = useGetFolder(id, searchKeyword, folderId);
+  const { link } = useGetFolderList(id, folderId);
   const { modalState, openModal } = useModal();
   const obsRef = useRef(null);
   const inputRef = useRef<HTMLInputElement>(null);
@@ -42,16 +46,33 @@ function Folder() {
 
   useEffect(() => {
     const observer = new IntersectionObserver(handleObserver);
-    if (!loading && obsRef.current) observer.observe(obsRef.current);
+    if (!loading && obsRef.current) {
+      observer.observe(obsRef.current);
+    }
     return () => {
       observer.disconnect();
     };
-  }, [loading]);
+  }, [loading, folderId]);
+
+  useEffect(() => {
+    for (let i = 0; i < link.length; i++) {
+      setWrongFolder(true);
+      if (link[i].id == folderId) {
+        setOnSelect({ id: folderId, name: link[i].name });
+        setWrongFolder(false);
+        break;
+      }
+    }
+  }, [link, folderId]);
+
+  if (folderId && wrongFolder) {
+    return <S.EmptyFolder>존재하지 않는 폴더입니다!</S.EmptyFolder>;
+  }
 
   return (
     <>
-      <div ref={obsRef}></div>
       <S.HeaderBody>
+        <div ref={obsRef}></div>
         <S.Header $view={toggleInput}>
           <S.HeaderModal>
             <S.LinkIcon src="/link.svg"></S.LinkIcon>
@@ -74,13 +95,13 @@ function Folder() {
         <SearchBar setSearchKeyWord={setSearchKeyWord} />
         {searchKeyword && (
           <S.SearchResult>
-            <p>{searchKeyword}</p>으로 검색한 결과입니다.
+            <p>{searchKeyword}</p>로 검색한 결과입니다.
           </S.SearchResult>
         )}
         <FolderButtonContainer link={link} setOnSelect={setOnSelect} />
         <S.FolderModalContainer>
-          {onSelect.name ? onSelect.name : '전체'}
-          {onSelect.name && (
+          {folderId ? onSelect.name : '전체'}
+          {folderId && (
             <FolderModals
               id={onSelect.id}
               name={onSelect.name}
